requirements.txt
defusedxml==0.4.1
Django==1.9.12
django-allauth==0.25.2
django-allauth-bootstrap==0.2
django-bootstrap-form==3.2.1
django-crispy-forms==1.6.1
django-social-auth==0.7.28
google-api-python-client==1.5.5
httplib2==0.9.2
oauth2==1.9.0.post1
oauth2client==4.0.0
oauthlib==2.0.1
Pillow==3.4.2
psycopg2==2.6.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyJWT==1.4.2
python-openid==2.2.5
python-social-auth==0.2.21
python3-openid==3.0.10
requests==2.12.4
requests-oauthlib==0.7.0
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
